requirements.txt
python-dotenv==0.21.1
selenium==4.24.0
pandas==1.5.3
bs4==0.0.1
